move pointinbbox helper to likewise utils file
